% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_2.0_functions.R
\name{assign_ids}
\alias{assign_ids}
\title{Assign New Ids}
\usage{
assign_ids(existing_ids, new_participants, prefix, digits = 3)
}
\arguments{
\item{existing_ids}{a vector of existing ids}

\item{new_participants}{a vector or new participants you want to create ids for}

\item{prefix}{a string prefix if required in id format  e.g. 'pch_' for pch_001}

\item{digits}{the number of digits in your id e.g. 3 for pch_001}
}
\value{
A numeric vector with the missing ids
}
\description{
this function will take existing ids and create new ones for new participants. It will fill missing ids.
}
