% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_2.0_functions.R
\name{calculate_itn}
\alias{calculate_itn}
\title{Calculates ITN}
\usage{
calculate_itn(data)
}
\arguments{
\item{data}{the data frame to act on}
}
\value{
A data frame with ITN
}
\description{
calculates ITN based on poverty threshold for number of people in household and annual income
}
