% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_2.0_functions.R
\name{calculate_timezone}
\alias{calculate_timezone}
\title{Calculate Timezone}
\usage{
calculate_timezone(zip_code)
}
\arguments{
\item{zip_code}{vector of zipcodes to act on (must be 5 digits)}
}
\value{
a vector of timezones for zipcodes
}
\description{
this function returns timezones for given zipcodes
}
