% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_2.0_functions.R
\name{city_info}
\alias{city_info}
\title{Pulling City Data}
\usage{
city_info()
}
\arguments{
\item{data}{data frame to act upon}
}
\value{
a dataframe with the race responses cleaned
}
\description{
this function pulls us city population data
}
