% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_2.0_functions.R
\name{clean_birthweight}
\alias{clean_birthweight}
\title{Standardizes birthweight responses}
\usage{
clean_birthweight(data)
}
\arguments{
\item{data}{the data frame to act on}
}
\value{
A data frame with the split birth weight columns (lb, oz)
}
\description{
takes character birthweight responses, cleans them and splits them into lb and oz columns
}
