% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_2.0_functions.R
\name{clean_race}
\alias{clean_race}
\title{Cleaning screener race responses}
\usage{
clean_race(data)
}
\arguments{
\item{data}{data frame to act upon}
}
\value{
a dataframe with the race responses cleaned
}
\description{
This function cleans race responses
}
