% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_2.0_functions.R
\name{flag_due_dates}
\alias{flag_due_dates}
\title{SUSPICIOUS DUE DATES}
\usage{
flag_due_dates(data)
}
\arguments{
\item{data}{The data frame you wish to act on}
}
\value{
A dataframe with a new column marking 1 for anybody with an impossible due date
}
\description{
Flags any response where the due date is over 40 weeks away or earlier than the day the screener was filled out
}
