% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_2.0_functions.R
\name{flag_duplicate_contacts}
\alias{flag_duplicate_contacts}
\title{Duplicate contacts}
\usage{
flag_duplicate_contacts(data)
}
\arguments{
\item{data}{The data frame you wish to act on}
}
\value{
A dataframe with 2 new columns marking 1 for anybody with a duplicate phone or email
}
\description{
Flags any participant with a duplicate phone or email
}
