% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_2.0_functions.R
\name{flag_ineligible_age}
\alias{flag_ineligible_age}
\title{INELIGIBLE AGES}
\usage{
flag_ineligible_age(data, threshold_date = Sys.Date() - 380)
}
\arguments{
\item{data}{The data frame you wish to act on}

\item{threshold_date}{The date threshold you wish to compare - default is 380 days prior to today}
}
\value{
A dataframe with a new column marking 1 for anybody too old to participate
}
\description{
Flags any child ages older than 380 days from today
}
