% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_2.0_functions.R
\name{flag_lowercase_names}
\alias{flag_lowercase_names}
\title{Lowercase Names}
\usage{
flag_lowercase_names(data)
}
\arguments{
\item{data}{The data frame you wish to act on}
}
\value{
A dataframe with a new column marking 1 for anybody with an all lowercase names
}
\description{
Flags any participant where their name is all lowercase
}
