% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_2.0_functions.R
\name{flag_numeric_names}
\alias{flag_numeric_names}
\title{Flagging Numeric Names}
\usage{
flag_numeric_names(data)
}
\arguments{
\item{data}{The data frame you wish to act on}
}
\value{
A dataframe with a new column marking 1 for anybody who has numeric values in their name
}
\description{
Flags any participant with numbers in their name
}
