% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_2.0_functions.R
\name{flag_referrals}
\alias{flag_referrals}
\title{Flagging Suspicious Referrals}
\usage{
flag_referrals(data, token, flagged_emails)
}
\arguments{
\item{data}{The data frame you wish to act on}

\item{token}{Unique REDCap API token}

\item{flagged_emails}{Data frame containing known flagged emails}
}
\value{
A dataframe with a new column marking 1 for anybody with flagged emails
}
\description{
Checks to see if snowball responses were referred by somebody we have flagged
}
