% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_2.0_functions.R
\name{get_all_snowball_responses}
\alias{get_all_snowball_responses}
\title{Pulling ALL Snowball Responses}
\usage{
get_all_snowball_responses(token, min_date_time = "2022-01-01 00:00:00")
}
\arguments{
\item{token}{Unique REDCap API token}

\item{min_date_time}{The minimum timestamp to pull responses after}
}
\value{
A list with 'snowball_responses': a df with new responses and the id that referred them, and 'ids_who_referred': a df with list of ids who have referred people, the number of people referred by them
}
\description{
Pulls new screener responses which have been referred by another person and all ids who have referred another person
}
