% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_2.0_functions.R
\name{get_ders}
\alias{get_ders}
\title{Process Difficulties in Emotion Regulation Data}
\usage{
get_ders(token, timepoint = "orca_4month_arm_1")
}
\arguments{
\item{token}{Unique REDCap token ID}

\item{timepoint}{redcap event name for the timepoint you wish to pull}
}
\value{
A data frame for the completed surveys
}
\description{
This function will download and return the total scores for each subscale of the DERS
}
