% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_2.0_functions.R
\name{get_eligible_responses}
\alias{get_eligible_responses}
\title{Pulling Eligible Screener Responses}
\usage{
get_eligible_responses(token, min_date_time = "2022-01-01 00:00:00")
}
\arguments{
\item{token}{Unique REDCap API token}

\item{min_date_time}{The minimum timestamp to pull responses after}
}
\value{
A data frame with screener record id and group participant was enrolled within (prenatal/postnatal)
}
\description{
Pulls screener responses who have been enrolled in the study
}
