% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_2.0_functions.R
\name{get_expected_invites}
\alias{get_expected_invites}
\title{Pull Expected Invites}
\usage{
get_expected_invites(token, timepoint = "4m", max_date = "none")
}
\arguments{
\item{token}{Unique REDCap token ID - use screener token if looking at timepoints 4m and below}

\item{timepoint}{'prenatal', 'newborn', '4m', '8m', '12m'}

\item{max_date}{Maximum date to cut off output - default is 'none'. Must be format of 'YYYY-MM-DD'}
}
\value{
A list with dataframe of all invites, bar chart, and dataframe with invite counts per month
}
\description{
This function will create a list with all future invites for a specified timepoint, a graph and counts showing invites per month
}
