% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_2.0_functions.R
\name{get_missing_ids}
\alias{get_missing_ids}
\title{Get Missing REDCap IDs}
\usage{
get_missing_ids(token)
}
\arguments{
\item{token}{Unique REDCap token ID}
}
\value{
A numeric vector with the missing ids
}
\description{
This function will pull any missing consecutive ids from redcap (that may need to be reassigned)
}
