% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_2.0_functions.R
\name{get_orca_bitsea}
\alias{get_orca_bitsea}
\title{Process BITSEA Data}
\usage{
get_orca_bitsea(token, timestamp = T)
}
\arguments{
\item{token}{Unique REDCap token ID}

\item{timestamp}{whether to include survey timestamp. Default is true}
}
\value{
A data frame for the completed surveys
}
\description{
This function will download and return the sum scores for the bitsea subscales
}
