% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_2.0_functions.R
\name{get_orca_cohort}
\alias{get_orca_cohort}
\title{Pull Records in Cohort}
\usage{
get_orca_cohort(token, screener = F)
}
\arguments{
\item{token}{Unique REDCap token ID}

\item{screener}{Set to 'T' to pull screener cohort (prospective/prenatal)}
}
\value{
A dataframe with all record ids included in cohort
}
\description{
This function will pull all records in a particular cohort (default is main ORCA postnatal)
}
