% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_2.0_functions.R
\name{get_orca_cws}
\alias{get_orca_cws}
\title{Process Cambridge Worry Scale Data}
\usage{
get_orca_cws(token, timestamp = T)
}
\arguments{
\item{token}{Unique REDCap token ID}

\item{timestamp}{whether to include survey timestamp. Default is true}
}
\value{
A data frame for the completed surveys
}
\description{
This function will download and return the mean scores for the cambridge worry scale
}
