% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_2.0_functions.R
\name{get_orca_data}
\alias{get_orca_data}
\title{Pulling ORCA Data}
\usage{
get_orca_data(
  token = token,
  form = form,
  raw_v_label = "raw",
  form_complete = T
)
}
\arguments{
\item{token}{The API token for the project}

\item{form}{The name of the REDCap form to retrieve data from}

\item{raw_v_label}{The label for raw data fields}

\item{form_complete}{Indicating whether you want to return all responses or just ones marked as complete (default is all)}
}
\value{
A data frame with the retrieved data
}
\description{
This function retrieves data from a REDCap project using the API.
}
