% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_2.0_functions.R
\name{get_orca_discrimination}
\alias{get_orca_discrimination}
\title{Process Discrimination Data}
\usage{
get_orca_discrimination(token, timestamp = T)
}
\arguments{
\item{token}{Unique REDCap token ID for New ORCA Recruitment Screener project}

\item{timestamp}{whether to include survey timestamp. Default is true}
}
\value{
A data frame for the completed surveys
}
\description{
This function will download and return the sum score for the discrimination q - not extra items
}
