% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_2.0_functions.R
\name{get_orca_epds}
\alias{get_orca_epds}
\title{Process EPDS postpartum depression data}
\usage{
get_orca_epds(token, timepoint = "orca_4month_arm_1")
}
\arguments{
\item{token}{Unique REDCap token ID}
}
\value{
A data frame for the completed surveys
}
\description{
This function will download and return the totalsummed scores for the EPDS. Only total scores, and not cutoff values, are returned.
}
