% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_2.0_functions.R
\name{get_orca_field}
\alias{get_orca_field}
\title{Pulling ORCA Field}
\usage{
get_orca_field(token = token, field = field, raw_v_label = "raw")
}
\arguments{
\item{token}{Unique REDCap token ID}

\item{field}{Name of the specific field to be downloaded}

\item{raw_v_label}{Whether raw data or labels are requested}
}
\value{
A data frame for the completed record_ids, redcap_event_name and field
}
\description{
PULLS RECORD ID AND INDIVIDUAL FIELD
}
