% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_2.0_functions.R
\name{get_orca_ibq}
\alias{get_orca_ibq}
\title{Process IBQ Data}
\usage{
get_orca_ibq(token)
}
\arguments{
\item{token}{Unique REDCap token ID}
}
\value{
A data frame for the completed surveys
}
\description{
This function will download and return the mean scores for the IBQ subscales (surgency, negative affect, effortful control).
}
