% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_2.0_functions.R
\name{get_orca_pbq}
\alias{get_orca_pbq}
\title{Process Postpartum Bonding Questionnaire Scores}
\usage{
get_orca_pbq(token, timepoint = "orca_4month_arm_1")
}
\arguments{
\item{token}{Unique REDCap token ID}
}
\value{
A data frame for the completed surveys
}
\description{
This function will download and return the total scores for pbq
}
