% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_2.0_functions.R
\name{get_orca_prenatal_surveys}
\alias{get_orca_prenatal_surveys}
\title{Pull all prenatal survey data}
\usage{
get_orca_prenatal_surveys(token)
}
\arguments{
\item{token}{Unique REDCap token ID}
}
\value{
A data frame for the completed surveys (order printed)
}
\description{
This function will download and return the mean scores for all ORCA prenatal surveys (will print order of surveys)
}
