% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_2.0_functions.R
\name{get_orca_report}
\alias{get_orca_report}
\title{Pulling a redcap report}
\usage{
get_orca_report(
  token,
  report_code,
  raw_or_label = "raw",
  checkbox_label = "false"
)
}
\arguments{
\item{token}{The API token for the project}

\item{report_code}{the character report code for the report you wish to pull}

\item{checkbox_label}{whether to combine checkbox fields}

\item{raw_v_label}{The label for raw data fields}
}
\value{
A dataframe with the report
}
\description{
Pulling a redcap report
}
