% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_2.0_functions.R
\name{get_orca_screener}
\alias{get_orca_screener}
\title{Pulling cleaned screener responses}
\usage{
get_orca_screener(token, min_date_time = "2022-01-01 00:00:00")
}
\arguments{
\item{token}{Unique REDCap API token}

\item{min_date_time}{The minimum timestamp to pull responses after}
}
\value{
a dataframe with the race responses cleaned
}
\description{
this function pulls and cleans screener data for import to excel tracker
}
