% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_2.0_functions.R
\name{get_orca_screener_clean}
\alias{get_orca_screener_clean}
\title{Get ORCA Screener Clean}
\usage{
get_orca_screener_clean(token, min_date_time = "2022-01-01 00:00:00")
}
\arguments{
\item{token}{Unique REDCap token ID for the New ORCA Recruitment Screener project}

\item{min_date_time}{The minimum timestamp to pull responses after}
}
\value{
A list containing data (screener to enroll), orca15 (potential reef people), prenatal_counts and postnatal_counts
}
\description{
Pulls new screener data, determines eligibility for enrollment
}
