% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_2.0_functions.R
\name{get_orca_social_support}
\alias{get_orca_social_support}
\title{Process Social Support Data}
\usage{
get_orca_social_support(token, timepoint = "orca_4month_arm_1", timestamp = T)
}
\arguments{
\item{token}{Unique REDCap token ID}

\item{timepoint}{Redcap event name for timepoint you wish to pull}

\item{timestamp}{whether to include survey timestamp. Default is true}
}
\value{
A data frame for the completed surveys
}
\description{
This function will download and return the sum and mean scores for the social support scale
}
