% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_2.0_functions.R
\name{get_prenatal_dob}
\alias{get_prenatal_dob}
\title{Pulling DOBs For Prenatal Participants}
\usage{
get_prenatal_dob(token)
}
\arguments{
\item{token}{Unique REDCap token ID}
}
\value{
A data frame with final dob and whether this is confirmed after birth or is just the due date
}
\description{
This function pulls all possible due date/dob update variables and returns most up to date
}
