% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_2.0_functions.R
\name{get_screener_child_dob}
\alias{get_screener_child_dob}
\title{Pull Screener Child DOB}
\usage{
get_screener_child_dob(token)
}
\arguments{
\item{token}{Unique REDCap token ID}
}
\value{
A data frame with all 3 dob variables (due date, rec_child_dob and child_dob_update) and final dob
}
\description{
This function will download and return the most up to date date of birth value from the screener
}
