% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_2.0_functions.R
\name{get_screener_contact_info}
\alias{get_screener_contact_info}
\title{Pulling Screener Contact Info}
\usage{
get_screener_contact_info(token)
}
\arguments{
\item{token}{Unique REDCap API token}
}
\value{
A data frame with screener record id, caregiver first/last name, phone number, texting preferences and email
}
\description{
Imports contact information from the screener
}
