% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_2.0_functions.R
\name{get_shipping_timetable}
\alias{get_shipping_timetable}
\title{Provides timepoint completion info for every timepoint}
\usage{
get_shipping_timetable(token)
}
\arguments{
\item{token}{Unique REDCap token ID}
}
\value{
A list with tracker (a df with every kit that needs to be shipped, and the date it needs to be shipped by) and counts (number of kits to be shipped each day)
}
\description{
Provides timepoint completion info for every timepoint
}
