% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_2.0_functions.R
\name{get_trimester_n}
\alias{get_trimester_n}
\title{Getting Trimester Numbers}
\usage{
get_trimester_n(token, field = "rec_due_date")
}
\arguments{
\item{token}{REDCap API token for SCREENER project}

\item{field}{string variable for name of your due date field, default is rec_due_date}
}
\value{
A dataframe with 'trimester' and 'n' column
}
\description{
This function will pull the number of prenatal participants in trimester 1, 2,3
}
