% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_2.0_functions.R
\name{get_visit_data}
\alias{get_visit_data}
\title{This function returns a dataframe with infant dob, date, and age for a specified event}
\usage{
get_visit_data(token, timepoint = 4)
}
\arguments{
\item{token}{Unique REDCap token ID}

\item{timepoint}{redcap_event_name for the timepoint you wish to pull}
}
\value{
dataframe with visit data for that timepoint
}
\description{
This function returns a dataframe with infant dob, date, and age for a specified event
}
