% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_2.0_functions.R
\name{get_visit_n}
\alias{get_visit_n}
\title{Generating N for Specified Timepoint}
\usage{
get_visit_n(token, timepoint = 4)
}
\arguments{
\item{token}{Unique REDCap token ID}

\item{timepoint}{integer reflecting the timepoint (4,8,12)}
}
\value{
Numeric variable for that event
}
\description{
This function with return a numeric variable for the number of people who have completed a particular timepoint
}
