% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_2.0_functions.R
\name{get_zip_info}
\alias{get_zip_info}
\title{Calculate all zipcode info}
\usage{
get_zip_info(zip_code, population_threshold = 50000)
}
\arguments{
\item{zip_code}{vector of zipcodes to act on (must be 5 digits)}

\item{population_threshold}{numeric value for population cut off for rural/non rural}
}
\value{
a dataframe with zipcode, location, timezone, population size and rurality
}
\description{
this function pulls location, timezone, population size and rurality for given zipcodes
}
