% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_2.0_functions.R
\name{import_data}
\alias{import_data}
\title{Import data to ORCA 2.0}
\usage{
import_data(token, data = data)
}
\arguments{
\item{token}{Unique REDCap API token}

\item{data}{the data frame you wish to import. Column names must match fields}
}
\value{
A summary of data to be imported/overwritten and requires user input (y) to continue with import
}
\description{
Imports the fields in a dataframe to the corresponding record IDS to redcap
}
