% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_2.0_functions.R
\name{import_mice_to_orca}
\alias{import_mice_to_orca}
\title{Importing MICE participant to 4m project}
\usage{
import_mice_to_orca(scr_token, orca_token, mice_record_id)
}
\arguments{
\item{scr_token}{Unique REDCap token ID for the New ORCA Recruitment Screener project}

\item{orca_token}{Unique REDCap token ID for the ORCA 2.0 project}

\item{mice_record_id}{String variable for the MICE id e.g. 'mc_018'}
}
\value{
User input for approval to import (respond y)
}
\description{
Pulls screener data for a mice id, reformats and imports it to the ORCA 2.0 redcap project
}
