% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_2.0_functions.R
\name{peach_enroll}
\alias{peach_enroll}
\title{PEACH Enrollment Function}
\usage{
peach_enroll(screener_token, peach_token, sri = "")
}
\arguments{
\item{screener_token}{Unique REDCap token ID for the New ORCA Recruitment Screener project}

\item{peach_token}{Unique REDCap token ID for the PEACH project}

\item{sri}{Screener record id for who you which to import}
}
\value{
A list containing data (screener to enroll), orca15 (potential reef people), prenatal_counts and postnatal_counts
}
\description{
For a given screener record id, it'll pull their peach participant info from screener, reformat and import to PEACH redcap. It'll also update pch_enroll_complete field on screener
}
