% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_2.0_functions.R
\name{replace_multiple}
\alias{replace_multiple}
\title{Replacing Multiple Values Function}
\usage{
replace_multiple(data, column_name, old_values, new_values)
}
\arguments{
\item{data}{The dataframe}

\item{column_name}{The column you wish to work on}

\item{old_values}{Vector of string variables you wish to replace}

\item{new_values}{Vector of string variables you wish to replace the old ones with}
}
\value{
the data frame with old strings in specified column replaced with new strings
}
\description{
ALLOWS DATA TO BE REPLACED IN A COLUMN
}
