% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_2.0_functions.R
\name{screen_fraudulence}
\alias{screen_fraudulence}
\title{Screens screener export}
\usage{
screen_fraudulence(data)
}
\arguments{
\item{data}{The data frame you wish to act on}
}
\value{
A list with the filtered screener, and a new column flagging lowercase names (doesn't remove). List also contains dataset with removed responses with duplicate contact info and ineligible ages
}
\description{
Checks for: duplicate contact info, under 18, numeric names, NA names/emails, babies above age threshold, names all lowercase
}
