% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_2.0_functions.R
\name{study_eligibility}
\alias{study_eligibility}
\title{Determine Study Eligibility}
\usage{
study_eligibility(data)
}
\arguments{
\item{data}{screener dataframe to act upon. Must contain 'child_dob' and 'priority' columns}
}
\value{
a dataframe with new column 'orca_study_enrollment' added
}
\description{
Takes a screener df and bins participants into different studies depending on age/demographic priority
}
