% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_2.0_functions.R
\name{zip_data}
\alias{zip_data}
\title{Pulls US zipcode database}
\usage{
zip_data()
}
\value{
A data frame with every us zipcode and information
}
\description{
Loads a data base of US zipcodes and information (latitude, longitude, population)
}
