% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_2.0_functions.R
\name{zips}
\alias{zips}
\title{ZIP CODE AND LOCATION FUNCTION}
\usage{
zips()
}
\value{
a dataframe with the race responses cleaned
}
\description{
This function pulls public US zipcode data and cleans it
}
